+++ package.json
@@ -12,7 +12,8 @@
     "bootstrap": "^5.3.2",
     "next": "14.0.4",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-loading-skeleton": "^3.3.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
